+++ indico/modules/events/registration/badges.py
@@ -102,6 +102,8 @@ class RegistrantsListToBadgesPDF(DesignerPDFBase):
             if placeholder:
                 if placeholder.group == 'registrant':
                     text = placeholder.render(registration)
+                elif placeholder.group == 'registration':
+                    text = placeholder.render(item['text'],registration.regform, registration)
                 else:
                     text = placeholder.render(registration.event)
             elif item['text']: